Create compiling_pest_on_max_osx
# Compilng `PEST` on `mac` `OSX
`
PEST (Parameter Estimation for any model) is probably the most widely used software to calibrate models of any kind. It has been used successfully to calibrate models of any kind and complexity all over the world. And its developer John Doherty has been improving, extending and developing in depth ever since the early 1990s. It is included in most user graphical user interfaces for groundwater and other type of models. PEST is now freele available on Windows as executable and UNIX as source from `http://www.pesthomepage.org/`, but it's lacking on OSX. A logical option is to compile the software from the UNIX source. Howeve that turned out less easy than it seemed. I sat together with John for a few hours in November 2016, trying to compile it, but was only very partially sucessful. Moreover, there is no information on the web telling how to do it.
I'v now solved the compilation on OSX problem completely and cleary, which was two days work. The how-to and why are in the uploaded file. Use it together with the pdf file that John provides to assist compilation  on UNIX.
I hope this helps a lot of people with Macs.

Theo Olsthoorn
Heemsted, Netherlands
2018-07-11 
Compiling PEST on OSX (mac) from the source of the UNIX version downloaded from
the PEST mainpage.